update: readme correct: compile + scripts feature: test common js and module examples
