+++ examples/index.js
@@ -3,26 +3,36 @@
 \*===========================================================================*/
 
 // Initialize
-import * as store from "rocket-store";
-
-const rs = await store.Rocketstore({
-	data_storage_area: "./webapp",
-	data_format: rocketstore._FORMAT_JSON,
-});
+const rs = require("rocket-store").default;
 
 (async () => {
 	// Change storage area from default ( <tempdir>/rsdb )
-	console.log("POST a record:\n", await rs.post("cars", "Mercedes_Benz_GT_R", { owner: "Lisa Simpson" }));
+	const options = {
+		data_storage_area: "./webapp",
+		data_format: rs._FORMAT_JSON,
+	};
+
+	await rs.options(options);
+
+	console.log(
+		"POST a record:\n",
+		await rs.post("cars", "Mercedes_Benz_GT_R", { owner: "Lisa Simpson" })
+	);
+
 	console.log("GET a record:\n", await rs.get("cars", ""));
-	console.log("-----");
-	console.log("POST 3 records");
+
 	// Post 3 records
-	await rs.post("cars", "BMW_740li", { owner: "Greg Onslow" }, rs._ADD_AUTO_INC);
+	await rs.post(
+		"cars",
+		"BMW_740li",
+		{ owner: "Greg Onslow" },
+		rs._ADD_AUTO_INC
+	);
 	await rs.post("cars", "BMW_740li", { owner: "Sam Wise" }, rs._ADD_AUTO_INC);
 	await rs.post("cars", "BMW_740li", { owner: "Bill Bo" }, rs._ADD_AUTO_INC);
+
 	console.log("Get all records:\n", await rs.get("cars", "*"));
-	console.log("-----");
-	console.log("Mass instert");
+
 	// Mass instert
 	const dataset = {
 		Gregs_BMW_740li: { owner: "Greg Onslow" },
@@ -43,13 +53,17 @@ const rs = await store.Rocketstore({
 	if (promises.length > 0) await Promise.all(promises);
 
 	console.log("Get BMW's:\n", await rs.get("cars", "*BMW*"));
-	console.log("Get list ordered by alphabetically descending keys:\n", await rs.get("cars", "", rs._ORDER_DESC));
 
-	console.log("Delete all Mercedes's:\n", await rs.delete("cars", "*Mercedes*"));
-	console.log("-----");
-	console.log("Return all cars");
+	console.log(
+		"Get list ordered by alphabetically descending keys:\n",
+		await rs.get("cars", "", rs._ORDER_DESC)
+	);
+
+	console.log(
+		"Delete all Mercedes's:\n",
+		await rs.delete("cars", "*Mercedes*")
+	);
 	console.log(await rs.get("cars", "*"));
-	console.log("-----");
-	console.log("Delete all records");
-	console.log(await rs.delete());
+
+	await rs.delete();
 })();
