+++ tsconfig-base.json
@@ -19,7 +19,7 @@
 		"traceResolution": false,
 		"types": ["node", "jest"]
 	},
-	"compileOnSave": false,
+	"compileOnSave": true,
 	"exclude": ["node_modules", "dist"],
 	"include": ["src"]
 }
